requirements.txt
numpy
Pillow
timm
torch
torchvision
